move karma command into own package
