+++ src/main/java/com/fossgalaxy/pircbotx/commands/karma/Karma.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License along with
  * uc_pircbotx. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.fossgalaxy.pircbotx.types;
+package com.fossgalaxy.pircbotx.commands.karma;
 
 /**
  * Represents a "thing" and an integer score associated with it
