+++ src/main/java/com/fossgalaxy/pircbotx/data/db/DatabaseModule.java
@@ -2,6 +2,7 @@ package com.fossgalaxy.pircbotx.data.db;
 
 import com.fossgalaxy.pircbotx.commands.factoid.FactoidModel;
 import com.fossgalaxy.pircbotx.commands.joins.JoinModel;
+import com.fossgalaxy.pircbotx.commands.karma.KarmaModel;
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
 import org.slf4j.Logger;
