+++ src/main/resources/META-INF/services/com.fossgalaxy.pircbotx.modules.Module
@@ -6,7 +6,7 @@ com.fossgalaxy.pircbotx.commands.joins.JoinsCommand
 com.fossgalaxy.pircbotx.commands.factoid.FactoidCommand
 com.fossgalaxy.pircbotx.commands.lart.LartCommand
 com.fossgalaxy.pircbotx.commands.NickCommand
-com.fossgalaxy.pircbotx.commands.KarmaCommand
+com.fossgalaxy.pircbotx.commands.karma.KarmaCommand
 com.fossgalaxy.pircbotx.commands.GamesCommand
 com.fossgalaxy.pircbotx.commands.CompSciCommand
 com.fossgalaxy.pircbotx.commands.IRCCommands
